Separate components from pages
