+++ App.js
@@ -4,8 +4,8 @@ import {createDrawerNavigator} from '@react-navigation/drawer';
 import {Portal, useTheme} from 'react-native-paper';
 import {StatusBar} from 'react-native';
 import Feed from './src/Feed';
-import DrawerContent from './src/DrawerContent';
-import AppBar from './src/AppBar';
+import DrawerContent from './src/components/DrawerContent';
+import AppBar from './src/components/AppBar';
 import Explore from './src/Explore';
 import Settings from './src/Settings';
 import Search from './src/Search';
