+++ src/components/AppBar.js
@@ -1,7 +1,7 @@
 import {useNavigation} from '@react-navigation/native';
 import React from 'react';
 import {Appbar, useTheme} from 'react-native-paper';
-import Logo from '../static/logo.svg';
+import Logo from '../../static/logo.svg';
 
 const DrawerContent = () => {
   const {colors} = useTheme();
