+++ src/components/Comment.js
@@ -1,16 +1,12 @@
 import * as React from 'react';
-import {View} from 'react-native';
-import {Card, IconButton, Text, useTheme} from 'react-native-paper';
-import {useWindowDimensions, Linking, Image} from 'react-native';
+import {Card, useTheme} from 'react-native-paper';
+import {useWindowDimensions} from 'react-native';
 import RenderHtml from 'react-native-render-html';
-import {InAppBrowser} from 'react-native-inappbrowser-reborn';
-import ago from 's-ago';
-import s from '../styles/Comment.module.css';
-import filter from '../utils/filter';
+import s from '../../styles/Comment.module.css';
+import filter from '../../utils/filter';
 import linkifyHtml from 'linkify-html';
-import UserChip from './UserChip';
 import AutoImage from './AutoImage';
-import {GlobalContext} from '../App';
+import {GlobalContext} from '../../App';
 
 const Comment = React.memo(({comment}) => {
   const {colors} = useTheme();
